requirements.txt
streamlit
pandas
openai
python-dotenv
requests
PyMuPDF